Allow files of domains to be specified as argument
Previously, the way to instruct the application to use multiple domains on the command-line was through what would appear a priori to be an unnecessary cat, but actually running `hstspreload batch` with a filename argument would cause the application to never halt. This sets it up so that files containing newline-separated domains can be used as arguments.